refactor: split ThingCLI CLI entrypoint from package API
