+++ tsconfig.json
@@ -5,6 +5,7 @@
         "noImplicitAny": false,
         "module": "CommonJS",
         "moduleResolution": "node",
+        "declaration": true,
         "esModuleInterop": true,
         "forceConsistentCasingInFileNames": true,
         "strict": true,
